+++ CHANGES.rst
@@ -12,6 +12,10 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 - Restore WebDAV support in Zope
   (`#744 <https://github.com/zopefoundation/Zope/issues/744>`_)
 
+- Fix longstanding test bug by forcing the page template engine.
+  Many tests in ``Products.PageTemplates`` used the old Zope page template
+  engine because the correct one was not registered during setup.
+
 - Remove deprecated module ``ZPublisher.maybe_lock``
   (`#758 <https://github.com/zopefoundation/Zope/issues/758>`_)